Escape error immediately before putting in html
Previously, we were using cgi.escape to convert html entities
into &lt;, &gt;, etc before sending the error message as a query
param. However, This causes a problem when the error message
contains *any* html entity.

Consider an error like: "'The <head> and the Heart' is not in your
watchlist". cgi.escape'ing this produces "'The &lt;head&gt; and the
Heart' is not in your watchlist". Trying to pass that as a query
parameter looks like:

    ?error=%27The%20&lt;...

The value of 'error' will only be "'The ", as the & indicates the start
of a new query parameter.

Another solution would be to urlencode the error message before passing
it as a parameter. Maybe we should do that anyway, as this solution
still suffers from being unable to cross of a movie like "Me, Myself &
Irene". Maybe we wanted to introduce that type of escaping later though?

In any case, it's a better practice to do html-escaping immediately
before putting the variable in HTML.